nfsconf|idmapd: fix segfault on musl libc systems.
Fixes: https://bugzilla.linux-nfs.org/show_bug.cgi?id=344

Symptoms: rpc.idmapd crashes upon being started and a
"general protection fault" is logged (ex: to /var/log/messages
on the system it was encountered on).

In summary, the problem was caused by an assumption that the
POSIX function sysconf(_SC_SOME_INDEX) would always return a
positive number, when it can actually return -1. On systems
that DO return positive values, the values are likely to be
guesses, which could lead to buffer overflow bugs.

The sysconf function was used to know how much memory to
allocate for obtaining user/group information from the
host system's passwd and group files/tables using
functions such as getpwnam_r and getgrnam_r.

This commit introduces a couple new structs and sets of
functions (ex: nfsutil_passwd_query, buffered_query,
nfsutil_pw_query_call_getpwnam_r) for managing calls to
those functions (ex: getpwnam_r and getgrnam_r) so that
these can be used in a way that does not require using the
sysconf function. Avoiding sysconf and fixed-size buffers
involves some additional memory management complexity.
Relatedly, one of my goals for these structs+functions is
to avoid excessive memory management code at the original
call sites. I worked towards this partially because I feel
that this helps the code remain readable, but also because
it greatly reduces the chances of making mistakes when
fixing analogous code that lives on codepaths that I can't
test or subject myself to.

This commit fixes the segfault specifically for the code
that loads the Nobody user+groups, as well as the code
that loads Static user and group mappings. However,
it does not apply the new structs/functions to the other
analogous codepaths that also use sysconf in the same manner.
I intend to apply a similar methodology to the other
codepaths later on, and offer a separate commit for that,
now that this fix seems to be working alright.

Error handling on the affected codepaths is now also more
comprehensive: the static.c functions now check for ALL
error codes that getpwnam_r and getgrnam_r return, while
the Nobody-handling code in libnfsidmap.c now distinguishes
between error-based failures and entries not being found
in passwd/group (and, relatedly, it is no longer possible
for this logic to pass 0 to strerror).