idmapd: Have regex.c use new passwd/group query code.
This should prevent any variants of bug-344 from happening in
this plugin.

It should also provide comprehensive error handling for
any get**name_r and get***id_r activity.

This commit also fixes a possible memory leak:
`regex_getpwnam` would not free the `localname` string if it
completed successfully. All of the error cases were covered,
just not success. That is now fixed, and it is written in
a way that makes it clear that the free() will always happen
if-and-only-if malloc() happened.
(It will also attempt to use a fixed-size stack allocated
buffer whenever the string fits.)