idmapd: Have nss.c use new passwd/group query code.
This should prevent any variants of bug-344 from happening
in this plugin.

This also brings with it the error handling/printing
improvements as mentioned in the previous commits in this
series, as well as some refactoring of other memory
management (string allocation) to make it clear that
leaks can't happen (this only applies to code that was
affected by passwd/group querying change).