idmapd: Have gums.c use new passwd/group query code.
This should prevent any variants of bug-344 from happening
in this plugin.

This also brings with it the error handling/printing
improvements as mentioned in the previous commits in this
series. In the case of this plugin, this applies to the
`translate_to_uid` and `translate_to_gid` functions;
other than that, there wasn't much to change in this
plugin (to make this bugfix/improvement work).

This plugin's `translate_to_uid` and `translate_to_gid`
will now retry-loop when get**nam_r returns EINTR
(retry on interrupt).

`translate_to_gid` used to have a realloc-loop, while
`translate_to_uid` did not. Now they both delegate the
realloc-looping to code in passwd_query.c, which should
make them equally robust.